+++ doc/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 Version 0.6
 -----------
 
-* add ``--stats`` command line option to TextRunner
+* add ``--hist`` and ``--stats`` command line options to TextRunner
 * :meth:`Benchmark.add_run` now raises an exception if a sample is zero.